fix: Eliminate all compiler warnings
Fixed warnings in multiple files:
- parser.cpp: Removed unused 'hasColon' variable, marked unused 'filename' params
- conditional.cpp: Marked unused 'lineNumber' parameters in processELSE/processENDIF
- expression.cpp: Removed unused 'startPos' variable in character literal parsing
- macro.cpp: Marked unused 'label' parameter in generateLocalLabel
- rel_writer.cpp: Fixed signed/unsigned comparison (int→size_t in loop)

All warnings eliminated while maintaining functionality.
Zero warnings in clean build.

+++ src/assembler/rel_writer.cpp
@@ -330,7 +330,7 @@ void RELWriter::writeSymbolName(const std::string& name) {
     std::string truncated = truncateSymbolName(name);
     
     // Pad to 6 characters with spaces
-    for (int i = 0; i < 6; i++) {
+    for (size_t i = 0; i < 6; i++) {
         char c = (i < truncated.length()) ? truncated[i] : ' ';
         writer_.writeByte(static_cast<Byte>(c));
     }
